refactor and new features
