+++ nextjs/app/components/Footer.jsx
@@ -4,9 +4,9 @@ import React from 'react'
 const Footer = () => {
     return (
         <div>
-            <p className='pl-2 flex flex-row'>Al reservar un turno, estas de acuedo con nuestras
+            <p className='pl-2 flex flex-row'>Al reservar y confirmar un turno, estás de acuedo con nuestra
                 <Link className='underline pl-1'
-                    href={'/politicasdereserva'}> politicas de reserva
+                    href={'/politicasdereserva'}> política de privacidad
                 </Link>
                 .</p>
         </div>
